+++ appDrawer/src/main/AndroidManifest.xml
