+++ demo/build.gradle
@@ -69,7 +69,7 @@ dependencies {
 
     //region Apps
     implementation project(path: ":apps")
-    implementation project(path: ":app_drawer")
+    implementation project(path: ":appDrawer")
     implementation project(path: ":app_search")
     implementation project(path: ":browser")
     implementation project(path: ":calculator")
