+++ demo/src/main/java/com/louis993546/metro/demo/MainActivity.kt
@@ -43,7 +43,7 @@ import androidx.navigation.compose.rememberNavController
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.HorizontalPager
 import com.google.accompanist.pager.rememberPagerState
-import com.louis993546.app_drawer.DrawerPage
+import com.louis993546.appDrawer.DrawerPage
 import com.louis993546.app_search.AppSearch
 import com.louis993546.apps.Apps
 import com.louis993546.calculator.CalculatorApp
