+++ settings.gradle
@@ -41,7 +41,7 @@ include ":radio"
 include ":calendar"
 include ":vertical_tiles_grid"
 include ":app_search"
-include ":app_drawer"
+include ":appDrawer"
 include ":launcher"
 include ":apps"
 include ":appRow"
